+++ src/components/@common/Tag/index.tsx
@@ -1,8 +1,3 @@
-/**
- * 태그 컴포넌트
- * @author 선우
- */
-
 import * as Styled from './Tag.styled';
 
 export interface ITagProps {
