+++ src/config/queryClient.ts
@@ -1,9 +1,10 @@
-import { QueryClient } from "@tanstack/react-query";
+import { QueryClient } from '@tanstack/react-query';
 
 export const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       retry: 1,
+      refetchOnWindowFocus: false,
     },
   },
 });
