+++ src/main.tsx
@@ -3,12 +3,12 @@ import { createRoot } from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
 import AppRouter from './routes/AppRouter.tsx';
 import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from '@/apis/@core/queryClient.ts';
+import { queryClient } from '@/config/queryClient.ts';
 import { ThemeProvider } from 'styled-components';
 import { Theme } from './styles/theme.ts';
 import GlobalStyle from './styles/GlobalStyle.ts';
 import { CookiesProvider } from 'react-cookie';
-// import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
@@ -19,7 +19,7 @@ createRoot(document.getElementById('root')!).render(
             <GlobalStyle />
             <AppRouter />
           </ThemeProvider>
-          {/* <ReactQueryDevtools initialIsOpen={false} /> */}
+          <ReactQueryDevtools initialIsOpen={false} />
         </QueryClientProvider>
       </RecoilRoot>
     </CookiesProvider>
